@@ -1,57 +1,66 @@
 {
-    "name": "sakai-ng",
-    "version": "20.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "watch": "ng build --watch --configuration development",
-        "format": "prettier --write \"**/*.{js,mjs,ts,mts,d.ts,html}\" --cache",
-        "test": "ng test"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^20",
-        "@angular/cdk": "^20.2.0",
-        "@angular/common": "^20",
-        "@angular/compiler": "^20",
-        "@angular/core": "^20",
-        "@angular/forms": "^20",
-        "@angular/platform-browser": "^20",
-        "@angular/platform-browser-dynamic": "^20",
-        "@angular/router": "^20",
-        "@primeuix/themes": "^1.2.1",
-        "@tailwindcss/postcss": "^4.1.11",
-        "chart.js": "4.4.2",
-        "primeclt": "^0.1.5",
-        "primeicons": "^7.0.0",
-        "primeng": "^20",
-        "rxjs": "~7.8.2",
-        "tailwindcss-primeui": "^0.6.1",
-        "tslib": "^2.8.1",
-        "zone.js": "~0.15.1"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^20",
-        "@angular/cli": "^20",
-        "@angular/compiler-cli": "^20",
-        "@types/jasmine": "~5.1.0",
-        "autoprefixer": "^10.4.20",
-        "eslint": "^9.30.1",
-        "eslint-config-prettier": "^10.1.5",
-        "eslint-plugin-import": "^2.32.0",
-        "eslint-plugin-prefer-arrow": "^1.2.3",
-        "eslint-plugin-prettier": "^5.5.1",
-        "jasmine-core": "~5.8.0",
-        "karma": "~6.4.4",
-        "karma-chrome-launcher": "~3.2.0",
-        "karma-coverage": "~2.2.1",
-        "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "~2.1.0",
-        "postcss": "^8.5.6",
-        "prettier": "^3.6.2",
-        "tailwindcss": "^4.1.11",
-        "typescript": "~5.8.3"
-    },
-    "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
+  "name": "sakai-ng",
+  "version": "20.0.0",
+  "type": "module",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "prebuild": "npm run lint && npm run check-format",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+    "lint": "ng lint",
+    "lint:fix": "ng lint --fix",
+    "format": "prettier --write \"src/**/*.{ts,html,scss,sass}\"",
+    "check-format": "prettier --check \"src/**/*.{ts,html,scss,sass}\""
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^20",
+    "@angular/cdk": "^20.2.0",
+    "@angular/common": "^20",
+    "@angular/compiler": "^20",
+    "@angular/core": "^20",
+    "@angular/forms": "^20",
+    "@angular/platform-browser": "^20",
+    "@angular/platform-browser-dynamic": "^20",
+    "@angular/router": "^20",
+    "@primeuix/themes": "^1.2.1",
+    "@tailwindcss/postcss": "^4.1.11",
+    "chart.js": "4.4.2",
+    "primeclt": "^0.1.5",
+    "primeicons": "^7.0.0",
+    "primeng": "^20",
+    "rxjs": "~7.8.2",
+    "tailwindcss-primeui": "^0.6.1",
+    "tslib": "^2.8.1",
+    "zone.js": "~0.15.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^20",
+    "@angular-eslint/builder": "20.2.0",
+    "@angular/cli": "^20",
+    "@angular/compiler-cli": "^20",
+    "@eslint/js": "^9.33.0",
+    "@types/jasmine": "~5.1.0",
+    "angular-eslint": "20.2.0",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^9.33.0",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
+    "eslint-plugin-prettier": "^5.5.1",
+    "jasmine-core": "~5.8.0",
+    "karma": "~6.4.4",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.1",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "postcss": "^8.5.6",
+    "prettier": "^3.6.2",
+    "tailwindcss": "^4.1.11",
+    "typescript": "~5.8.3",
+    "typescript-eslint": "8.40.0"
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
