@@ -1,44 +1,50 @@
-// src/app/pages/cliente/cliente.component.ts
-
-import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
-import { ClienteService } from 'src/app/pages/service/cliente.service';
-import { ClienteRequest } from 'src/app/pages/models/cliente.model'; // Usa ClienteRequest
+import { Component } from '@angular/core';
+import {
+  FormControl,
+  FormGroup,
+  FormsModule,
+  ReactiveFormsModule,
+  Validators,
+} from '@angular/forms';
+import { Toast } from 'primeng/toast';
 import { MessageService } from 'primeng/api';
-import { ToastModule } from 'primeng/toast';
-import { ButtonModule } from 'primeng/button';
-import { InputTextModule } from 'primeng/inputtext';
+import { Button } from 'primeng/button';
 import { FloatLabelModule } from 'primeng/floatlabel';
-import { CommonModule } from '@angular/common';
+import { InputTextModule } from 'primeng/inputtext';
+import { ClientRequest } from '@/common/api/interfaces/requests/ClientRequest';
+import { Client } from '@/common/api/services/client';
 
 @Component({
-  selector: 'app-cliente',
-  standalone: true,
+  selector: 'app-client',
   imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    ToastModule,
-    ButtonModule,
+    Toast,
+    Button,
+    FormsModule,
+    FloatLabelModule,
     InputTextModule,
-    FloatLabelModule
+    ReactiveFormsModule,
   ],
-  // No necesitas providers: MessageService ya está en appConfig
+  providers: [MessageService],
   template: `
     <div class="card p-6 max-w-4xl mx-auto">
       <h2 class="text-2xl font-semibold mb-6">Crear Cliente</h2>
-      <p-toast></p-toast>
+      <p-toast />
 
-      <form [formGroup]="clienteForm" (ngSubmit)="guardarCliente()" class="grid grid-cols-1 md:grid-cols-2 gap-6">
-
+      <form
+        [formGroup]="clienteForm"
+        (ngSubmit)="guardarCliente()"
+        class="grid grid-cols-1 md:grid-cols-2 gap-6"
+      >
         <!-- Nombre (obligatorio) -->
         <div>
           <p-floatlabel>
             <input pInputText formControlName="name" id="name" />
             <label for="name">Nombre *</label>
           </p-floatlabel>
-          <small class="text-red-500" *ngIf="name.invalid && name.touched">
-            Nombre es obligatorio
-          </small>
+
+          @if (name.invalid && name.touched) {
+            <small class="text-red-500">Nombre es obligatorio</small>
+          }
         </div>
 
         <!-- Apellido -->
@@ -55,9 +61,10 @@ import { CommonModule } from '@angular/common';
             <input pInputText formControlName="address" id="address" />
             <label for="address">Dirección *</label>
           </p-floatlabel>
-          <small class="text-red-500" *ngIf="address.invalid && address.touched">
-            Dirección es obligatoria
-          </small>
+
+          @if (address.invalid && address.touched) {
+            <small class="text-red-500">Dirección es obligatoria</small>
+          }
         </div>
 
         <!-- Celular (obligatorio) -->
@@ -66,9 +73,10 @@ import { CommonModule } from '@angular/common';
             <input pInputText formControlName="phoneNumber" id="phoneNumber" />
             <label for="phoneNumber">Celular *</label>
           </p-floatlabel>
-          <small class="text-red-500" *ngIf="phoneNumber.invalid && phoneNumber.touched">
-            Celular es obligatorio
-          </small>
+
+          @if (phoneNumber.invalid && phoneNumber.touched) {
+            <small class="text-red-500">Celular es obligatorio</small>
+          }
         </div>
 
         <!-- Correo -->
@@ -97,67 +105,71 @@ import { CommonModule } from '@angular/common';
 
         <!-- Botón -->
         <div class="md:col-span-2 mt-4">
-          <p-button 
-            label="Guardar Cliente" 
-            icon="pi pi-save" 
-            [loading]="loading" 
+          <p-button
+            label="Guardar Cliente"
+            icon="pi pi-save"
+            [loading]="loading"
             type="submit"
             [disabled]="clienteForm.invalid"
           />
         </div>
       </form>
     </div>
-  `
+  `,
 })
-export class ClienteComponent implements OnInit {
-  clienteForm: FormGroup;
-  loading = false;
+export class ClientPage {
+  protected clienteForm = new FormGroup({
+    name: new FormControl('', [Validators.required]),
+    lastName: new FormControl('', [Validators.required]),
+    address: new FormControl('', [Validators.required]),
+    phoneNumber: new FormControl('', [Validators.required]),
+    email: new FormControl('', [Validators.email]),
+    dni: new FormControl(''),
+    cuit: new FormControl(''),
+  });
+  protected loading = false;
 
   constructor(
-    private fb: FormBuilder,
-    private clienteService: ClienteService,
-    private messageService: MessageService
-  ) {
-    this.clienteForm = this.fb.group({
-      name: ['', [Validators.required]],
-      lastName: ['', [Validators.required]],
-      address: ['', [Validators.required]],
-      phoneNumber: ['', [Validators.required]],
-      email: ['', [Validators.email]],
-      dni: [''],
-      cuit: ['']
-    });
-  }
-
-  ngOnInit(): void {}
+    private clienteService: Client,
+    private messageService: MessageService,
+  ) {}
 
   // Getters para validación
-  get name() { return this.clienteForm.get('name')!; }
-  get lastName() { return this.clienteForm.get('lastName')!; }
-  get address() { return this.clienteForm.get('address')!; }
-  get phoneNumber() { return this.clienteForm.get('phoneNumber')!; }
+  get name() {
+    return this.clienteForm.get('name')!;
+  }
+  get lastName() {
+    return this.clienteForm.get('lastName')!;
+  }
+  get address() {
+    return this.clienteForm.get('address')!;
+  }
+  get phoneNumber() {
+    return this.clienteForm.get('phoneNumber')!;
+  }
 
-  guardarCliente() {
+  protected guardarCliente() {
     if (this.clienteForm.invalid) {
       this.messageService.add({
         severity: 'warn',
         summary: 'Faltan datos',
-        detail: 'Nombre, apellido, dirección y celular son obligatorios'
+        detail: 'Nombre, apellido, dirección y celular son obligatorios',
       });
+
       return;
     }
 
     this.loading = true;
 
     // El objeto ya está en el formato correcto: name, lastName, phoneNumber, etc.
-    const cliente: ClienteRequest = this.clienteForm.value;
+    const cliente = this.clienteForm.value as ClientRequest;
 
-    this.clienteService.create(cliente).subscribe({
+    this.clienteService.createClient(cliente).subscribe({
       next: (response) => {
         this.messageService.add({
           severity: 'success',
           summary: 'Éxito',
-          detail: `Cliente creado con ID: ${response.clientId}`
+          detail: `Cliente creado con ID: ${response}`,
         });
         this.clienteForm.reset();
       },
@@ -165,12 +177,14 @@ export class ClienteComponent implements OnInit {
         this.messageService.add({
           severity: 'error',
           summary: 'Error',
-          detail: 'No se pudo guardar el cliente: ' + (err.message || 'Error de conexión')
+          detail:
+            'No se pudo guardar el cliente: ' +
+            (err.message || 'Error de conexión'),
         });
       },
       complete: () => {
         this.loading = false;
-      }
+      },
     });
   }
 }
