@@ -1,4 +1,4 @@
-import { environment } from "src/environments/environment";
+import { environment } from 'src/environments/environment';
 
 const API_BASE_URL = environment.backend_url;
 
@@ -27,4 +27,4 @@ export const Endpoints = {
   FIXED_COST: API_BASE_URL + '/FixedCost',
   PAYMENTS: API_BASE_URL + '/Payments',
   PERSONS: API_BASE_URL + '/Person',
-}
+};
