@@ -1,8 +1,8 @@
 export interface BudgetRequest {
-    budgetId?: number;
-    descriptionBudget: string;
-    clientId: number;
-    budgetStatus: string;
-    dateCreated: Date;
-    deadLine: Date;
+  budgetId?: number;
+  descriptionBudget: string;
+  clientId: number;
+  budgetStatus: string;
+  dateCreated: Date;
+  deadLine: Date;
 }
