@@ -1,5 +1,5 @@
 export interface CategoryRequest {
-    CategoryId: string;
-    CategoryName: string;
-    CategoryModel: string;
+  CategoryId: string;
+  CategoryName: string;
+  CategoryModel: string;
 }
