@@ -1,10 +1,10 @@
 export interface ClientRequest {
-    clientId?: number;
-    name: string;
-    lastName: string;
-    address: string;
-    phoneNumber: string;
-    email?: string;
-    dni?: string;
-    cuit?: string;
+  clientId?: number;
+  name: string;
+  lastName: string;
+  address: string;
+  phoneNumber: string;
+  email?: string;
+  dni?: string;
+  cuit?: string;
 }
