@@ -1,11 +1,11 @@
 export interface EmployeeRequest {
-    employeeId?: number;
-    salary?: number;
-    name: string;
-    lastName: string;
-    address: string;
-    phoneNumber: string;
-    email: string;
-    dni: string;
-    cuit: string;
+  employeeId?: number;
+  salary?: number;
+  name: string;
+  lastName: string;
+  address: string;
+  phoneNumber: string;
+  email: string;
+  dni: string;
+  cuit: string;
 }
