@@ -1,7 +1,7 @@
 export interface FixedCostRequest {
-    Description: string;
-    Amount: number;
-    WorkingDays: number;
-    HoursWorked: number;
-    DateCharged: Date;
+  Description: string;
+  Amount: number;
+  WorkingDays: number;
+  HoursWorked: number;
+  DateCharged: Date;
 }
