@@ -1,7 +1,7 @@
 export interface InvoiceItemRequest {
-    itemId?: number;
-    MaterialId: number;
-    Quantity: number;
-    Price: number;
-    InvoiceId: number;
+  itemId?: number;
+  MaterialId: number;
+  Quantity: number;
+  Price: number;
+  InvoiceId: number;
 }
