@@ -1,6 +1,6 @@
 export interface InvoiceRequest {
-    date: Date;
-    isPaid: boolean;
-    supplierId: number;
-    invoiceId?: number;
+  date: Date;
+  isPaid: boolean;
+  supplierId: number;
+  invoiceId?: number;
 }
