@@ -1,10 +1,10 @@
 export interface MaterialRequest {
-    MaterialId?: number;
-    MaterialName: string;
-    MaterialDescription: string;
-    MaterialColor: string;
-    MaterialBrand: string;
-    MaterialMeasure: string;
-    MaterialUnitMeasure: string;
-    SubCategoryMaterialId: number;
+  MaterialId?: number;
+  MaterialName: string;
+  MaterialDescription: string;
+  MaterialColor: string;
+  MaterialBrand: string;
+  MaterialMeasure: string;
+  MaterialUnitMeasure: string;
+  SubCategoryMaterialId: number;
 }
