@@ -1,9 +1,9 @@
 export interface PersonRequest {
-    name: string;
-    lastName: string;
-    address: string;
-    phoneNumber: string;
-    email: string;
-    dni: string;
-    cuit: string;
+  name: string;
+  lastName: string;
+  address: string;
+  phoneNumber: string;
+  email: string;
+  dni: string;
+  cuit: string;
 }
