@@ -1,11 +1,11 @@
 export interface SupplierRequest {
-    supplierId?: number;
-    note?: string;
-    name: string;
-    lastName: string;
-    address: string;
-    phoneNumber: string;
-    email: string;
-    dni: string;
-    cuit: string;
+  supplierId?: number;
+  note?: string;
+  name: string;
+  lastName: string;
+  address: string;
+  phoneNumber: string;
+  email: string;
+  dni: string;
+  cuit: string;
 }
