@@ -1,5 +1,5 @@
 export interface CategoryResponse {
-    categoryId: number;
-    categoryName: string;
-    categoryModel: string;
+  categoryId: number;
+  categoryName: string;
+  categoryModel: string;
 }
