@@ -2,7 +2,7 @@ import { BudgetResponse } from './BudgetResponse';
 import { ClientResponse } from './ClientResponse';
 
 export interface ClientHistoryResponse {
-    clientHistoryId: number;
-    clientId: ClientResponse;
-    budgetsId: BudgetResponse[];
+  clientHistoryId: number;
+  clientId: ClientResponse;
+  budgetsId: BudgetResponse[];
 }
