@@ -1,6 +1,6 @@
 import { PersonResponse } from './PersonResponse';
 
 export interface ClientResponse {
-    clientId: number;
-    personId: PersonResponse;
+  clientId: number;
+  personId: PersonResponse;
 }
