@@ -1,6 +1,5 @@
 export interface CostResponse {
-
-    CostId: number;
-    CostValue: number;
-    Description: string;
+  CostId: number;
+  CostValue: number;
+  Description: string;
 }
