@@ -2,7 +2,7 @@ import { EmployeeResponse } from './EmployeeResponse';
 import { SalaryResponse } from './SalaryResponse';
 
 export interface EmployeeHistoryResponse {
-    idEmployeeHistory: number;
-    oEmployee: EmployeeResponse;
-    salaries: SalaryResponse[];
+  idEmployeeHistory: number;
+  oEmployee: EmployeeResponse;
+  salaries: SalaryResponse[];
 }
