@@ -1,7 +1,7 @@
 import { PersonResponse } from './PersonResponse';
 
 export interface EmployeeResponse {
-    idEmployee: number;
-    oPerson: PersonResponse;
-    salary?: number;
+  idEmployee: number;
+  oPerson: PersonResponse;
+  salary?: number;
 }
