@@ -1,8 +1,8 @@
 export interface FixedCostResponse {
-    idFixedCost: number;
-    description: string;
-    amount: number;
-    workingDays: number;
-    hoursWorked: number;
-    date: Date;
+  idFixedCost: number;
+  description: string;
+  amount: number;
+  workingDays: number;
+  hoursWorked: number;
+  date: Date;
 }
