@@ -1,9 +1,9 @@
 import { MaterialResponse } from './MaterialResponse';
 
 export interface InvoiceItemResponse {
-    invoiceId?: number;
-    invoiceItemId: number;
-    oMaterial: MaterialResponse;
-    quantity: number;
-    price: number;
+  invoiceId?: number;
+  invoiceItemId: number;
+  oMaterial: MaterialResponse;
+  quantity: number;
+  price: number;
 }
