@@ -1,7 +1,7 @@
 import { MaterialResponse } from './MaterialResponse';
 
 export interface ItemResponse {
-    itemId: number;
-    oMaterial: MaterialResponse;
-    quantity: number;
+  itemId: number;
+  oMaterial: MaterialResponse;
+  quantity: number;
 }
