@@ -1,6 +1,6 @@
 export interface PaymentResponse {
-    idPayment: number;
-    date: Date;
-    amount: number;
-    description: string;
+  idPayment: number;
+  date: Date;
+  amount: number;
+  description: string;
 }
