@@ -1,10 +1,10 @@
 export interface PersonResponse {
-    personId: number;
-    name: string;
-    lastName: string;
-    address: string;
-    phoneNumber: string;
-    email: string;
-    dni: string;
-    cuit: string;
+  personId: number;
+  name: string;
+  lastName: string;
+  address: string;
+  phoneNumber: string;
+  email: string;
+  dni: string;
+  cuit: string;
 }
