@@ -1,8 +1,8 @@
 import { PaymentResponse } from './PaymentResponse';
 
 export interface SalaryResponse {
-    idSalary: number;
-    amount: number;
-    billDate: number;
-    payments: PaymentResponse[];
+  idSalary: number;
+  amount: number;
+  billDate: number;
+  payments: PaymentResponse[];
 }
