@@ -1,7 +1,7 @@
 import { CategoryResponse } from './CategoryResponse';
 
 export interface SubCategoryMaterialResponse {
-    subCategoryMaterialId: number;
-    subCategoryName: string;
-    categoryId: CategoryResponse;
+  subCategoryMaterialId: number;
+  subCategoryName: string;
+  categoryId: CategoryResponse;
 }
