@@ -1,7 +1,7 @@
 import { PersonResponse } from './PersonResponse';
 
 export interface SupplierResponse {
-    supplierId: number;
-    note?: string;
-    personId: PersonResponse;
+  supplierId: number;
+  note?: string;
+  personId: PersonResponse;
 }
