@@ -5,23 +5,17 @@ import { BudgetResponse } from '../interfaces/responses/BudgetResponse';
 import { BudgetRequest } from '../interfaces/requests/BudgetRequest';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class Budget {
-  constructor(
-    private http: HttpClient,
-  ) {}
+  constructor(private http: HttpClient) {}
 
   public getBudgets() {
-    return this.http.get<BudgetResponse[]>(
-      Endpoints.BUDGET,
-    );
+    return this.http.get<BudgetResponse[]>(Endpoints.BUDGET);
   }
 
   public getBudgetById(id_budget: string) {
-    return this.http.get<BudgetResponse>(
-      `${Endpoints.BUDGET}/${id_budget}`,
-    );
+    return this.http.get<BudgetResponse>(`${Endpoints.BUDGET}/${id_budget}`);
   }
 
   public getBudgetByClientId(id_client: string) {
@@ -31,35 +25,24 @@ export class Budget {
   }
 
   public getBudgetCost(id_budget: string) {
-    return this.http.get<number>(
-      `${Endpoints.BUDGET_PRICE}/${id_budget}`,
-    );
+    return this.http.get<number>(`${Endpoints.BUDGET_PRICE}/${id_budget}`);
   }
 
   public createBudget(data: BudgetRequest) {
-    return this.http.post<void>(
-      Endpoints.BUDGET,
-      data,
-    );
-  }
-  
-  public updateBudget(data: BudgetRequest) {
-    return this.http.put<void>(
-      Endpoints.BUDGET,
-      data,
-    );
-  }
-  
-  public deleteBudget(id_budget: number) {
-    return this.http.delete<void>(
-      `${Endpoints.BUDGET}/${id_budget}`,
-    );
+    return this.http.post<void>(Endpoints.BUDGET, data);
   }
 
-  public CalculateBudgetPrice(id_budget: string) {
+  public updateBudget(data: BudgetRequest) {
+    return this.http.put<void>(Endpoints.BUDGET, data);
+  }
+
+  public deleteBudget(id_budget: number) {
+    return this.http.delete<void>(`${Endpoints.BUDGET}/${id_budget}`);
+  }
+
+  public calculateBudgetPrice(id_budget: string) {
     return this.http.get<number>(
       `${Endpoints.BUDGET_PRICE}/${id_budget}`,
     );
   }
-
 }
