@@ -2,21 +2,16 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Endpoints } from '../constants/Endpoints';
 import { CategoryRequest } from '../interfaces/requests/CategoryRequest';
-import{ CategoryResponse } from '../interfaces/responses/CategoryResponse';
+import { CategoryResponse } from '../interfaces/responses/CategoryResponse';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class Category {
-
-    constructor(
-    private http: HttpClient,
-  ) {}
+  constructor(private http: HttpClient) {}
 
   public getCategories() {
-    return this.http.get<CategoryResponse[]>(
-      Endpoints.CATEGORIES,
-    );
+    return this.http.get<CategoryResponse[]>(Endpoints.CATEGORIES);
   }
 
   public getCategoryById(id_category: string) {
@@ -26,23 +21,14 @@ export class Category {
   }
 
   public createCategory(data: CategoryRequest) {
-    return this.http.post<void>(
-      Endpoints.CATEGORIES,
-      data,
-    );
+    return this.http.post<void>(Endpoints.CATEGORIES, data);
   }
-  
+
   public updateCategory(data: CategoryRequest) {
-    return this.http.put<void>(
-      Endpoints.CATEGORIES,
-      data,
-    );
+    return this.http.put<void>(Endpoints.CATEGORIES, data);
   }
-  
+
   public deleteCategory(id_category: number) {
-    return this.http.delete<void>(
-      `${Endpoints.CATEGORIES}/${id_category}`,
-    );
+    return this.http.delete<void>(`${Endpoints.CATEGORIES}/${id_category}`);
   }
-  
 }
