@@ -5,25 +5,44 @@ import { LayoutService } from '../service/layout.service';
 import { CommonModule } from '@angular/common';
 
 @Component({
-    selector: 'app-floating-configurator',
-    imports: [CommonModule, ButtonModule, StyleClassModule],
-    template: `
-        <div class="flex gap-4 top-8 right-8" [ngClass]="{ fixed: float() }">
-            <p-button type="button" (onClick)="toggleDarkMode()" [rounded]="true" [icon]="isDarkTheme() ? 'pi pi-moon' : 'pi pi-sun'" severity="secondary" />
-            <div class="relative">
-                <p-button icon="pi pi-palette" pStyleClass="@next" enterFromClass="hidden" enterActiveClass="animate-scalein" leaveToClass="hidden" leaveActiveClass="animate-fadeout" [hideOnOutsideClick]="true" type="button" rounded />
-            </div>
-        </div>
-    `
+  selector: 'app-floating-configurator',
+  imports: [CommonModule, ButtonModule, StyleClassModule],
+  template: `
+    <div class="flex gap-4 top-8 right-8" [ngClass]="{ fixed: float() }">
+      <p-button
+        type="button"
+        (onClick)="toggleDarkMode()"
+        [rounded]="true"
+        [icon]="isDarkTheme() ? 'pi pi-moon' : 'pi pi-sun'"
+        severity="secondary"
+      />
+      <div class="relative">
+        <p-button
+          icon="pi pi-palette"
+          pStyleClass="@next"
+          enterFromClass="hidden"
+          enterActiveClass="animate-scalein"
+          leaveToClass="hidden"
+          leaveActiveClass="animate-fadeout"
+          [hideOnOutsideClick]="true"
+          type="button"
+          rounded
+        />
+      </div>
+    </div>
+  `,
 })
 export class AppFloatingConfigurator {
-    LayoutService = inject(LayoutService);
+  LayoutService = inject(LayoutService);
 
-    float = input<boolean>(true);
+  float = input<boolean>(true);
 
-    isDarkTheme = computed(() => this.LayoutService.layoutConfig().darkTheme);
+  isDarkTheme = computed(() => this.LayoutService.layoutConfig().darkTheme);
 
-    toggleDarkMode() {
-        this.LayoutService.layoutConfig.update((state) => ({ ...state, darkTheme: !state.darkTheme }));
-    }
+  toggleDarkMode() {
+    this.LayoutService.layoutConfig.update((state) => ({
+      ...state,
+      darkTheme: !state.darkTheme,
+    }));
+  }
 }
