@@ -2,13 +2,15 @@ import { Component, ElementRef } from '@angular/core';
 import { AppMenu } from './app.menu';
 
 @Component({
-    selector: 'app-sidebar',
-    standalone: true,
-    imports: [AppMenu],
-    template: ` <div class="layout-sidebar">
-        <app-menu></app-menu>
-    </div>`
+  selector: 'app-sidebar',
+  standalone: true,
+  imports: [AppMenu],
+  template: `
+    <div class="layout-sidebar">
+      <app-menu />
+    </div>
+  `,
 })
 export class AppSidebar {
-    constructor(public el: ElementRef) {}
+  constructor(public el: ElementRef) {}
 }
