@@ -4,7 +4,7 @@ import { Login } from './login';
 import { Error } from './error';
 
 export default [
-    { path: 'access', component: Access },
-    { path: 'error', component: Error },
-    { path: 'login', component: Login }
+  { path: 'access', component: Access },
+  { path: 'error', component: Error },
+  { path: 'login', component: Login },
 ] as Routes;
