@@ -1,84 +1,55 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { CardWidget } from './cardwidget';
 
 @Component({
-    standalone: true,
-    selector: 'app-stats-widget',
-    imports: [CommonModule],
-    template: `<div class="col-span-12 lg:col-span-6 xl:col-span-6">
-            <div class="card mb-0">
-                <div class="flex justify-between mb-4">
-                    <div>
-                        <span class="block text-muted-color font-medium mb-4">Presupuestos</span>
-                        <div class="text-surface-900 dark:text-surface-0 font-medium text-xl">152</div>
-                    </div>
-                    <div class="flex items-center justify-center bg-blue-100 dark:bg-blue-400/10 rounded-border" style="width: 2.5rem; height: 2.5rem">
-                        <i class="pi pi-file-edit text-blue-500 text-xl!"></i>
-                    </div>
-                </div>
-                <span class="text-primary font-medium">24 nuevos </span>
-                <span class="text-muted-color">desde la ultima visita</span>
-            </div>
-        </div>
-        <div class="col-span-12 lg:col-span-6 xl:col-span-6">
-            <div class="card mb-0">
-                <div class="flex justify-between mb-4">
-                    <div>
-                        <span class="block text-muted-color font-medium mb-4">Trabajos</span>
-                        <div class="text-surface-900 dark:text-surface-0 font-medium text-xl">120 Realizados</div>
-                    </div>
-                    <div class="flex items-center justify-center bg-orange-100 dark:bg-orange-400/10 rounded-border" style="width: 2.5rem; height: 2.5rem">
-                        <i class="pi pi-hammer text-orange-500 text-xl!"></i>
-                    </div>
-                </div>
-                <span class="text-primary font-medium">%52+ </span>
-                <span class="text-muted-color">desde la semana pasada</span>
-            </div>
-        </div>
-        <div class="col-span-12 lg:col-span-6 xl:col-span-4">
-            <div class="card mb-0">
-                <div class="flex justify-between mb-4">
-                    <div>
-                        <span class="block text-muted-color font-medium mb-4">Clientes</span>
-                        <div class="text-surface-900 dark:text-surface-0 font-medium text-xl">28441</div>
-                    </div>
-                    <div class="flex items-center justify-center bg-cyan-100 dark:bg-cyan-400/10 rounded-border" style="width: 2.5rem; height: 2.5rem">
-                        <i class="pi pi-users text-cyan-500 text-xl!"></i>
-                    </div>
-                </div>
-                <span class="text-primary font-medium">520 </span>
-                <span class="text-muted-color">registrados recientemente</span>
-            </div>
-        </div>
-        <div class="col-span-12 lg:col-span-6 xl:col-span-4">
-            <div class="card mb-0">
-                <div class="flex justify-between mb-4">
-                    <div>
-                        <span class="block text-muted-color font-medium mb-4">Materiales</span>
-                        <div class="text-surface-900 dark:text-surface-0 font-medium text-xl">152</div>
-                    </div>
-                    <div class="flex items-center justify-center bg-purple-100 dark:bg-purple-400/10 rounded-border" style="width: 2.5rem; height: 2.5rem">
-                        <i class="pi pi-box text-purple-500 text-xl!"></i>
-                    </div>
-                </div>
-                <span class="text-primary font-medium">85 </span>
-                <span class="text-muted-color">añadidos recientemente</span>
-            </div>
-        </div>
-        <div class="col-span-12 lg:col-span-6 xl:col-span-4">
-            <div class="card mb-0">
-                <div class="flex justify-between mb-4">
-                    <div>
-                        <span class="block text-muted-color font-medium mb-4">Proveedores</span>
-                        <div class="text-surface-900 dark:text-surface-0 font-medium text-xl">152</div>
-                    </div>
-                    <div class="flex items-center justify-center bg-purple-100 dark:bg-purple-400/10 rounded-border" style="width: 2.5rem; height: 2.5rem">
-                        <i class="pi pi-building text-yellow-500 text-xl!"></i>
-                    </div>
-                </div>
-                <span class="text-primary font-medium">20 </span>
-                <span class="text-muted-color">registrados recientemente</span>
-            </div>
-        </div>`
+  standalone: true,
+  selector: 'app-stats-widget',
+  imports: [CardWidget],
+  template: `
+    <app-card-widget
+      class="col-span-3 lg:col-span-2 xl:col-span-1"
+      title="Presupuestos"
+      description="152"
+      icon="pi-file-edit"
+      colorIcon="text-blue-500"
+      colorBoxIcon="bg-blue-100"
+    />
+
+    <app-card-widget
+      class="col-span-3 lg:col-span-2 xl:col-span-1"
+      title="Trabajos"
+      description="120 Realizados"
+      icon="pi-hammer"
+      colorIcon="text-orange-500"
+      colorBoxIcon="bg-orange-100"
+    />
+
+    <app-card-widget
+      class="col-span-3 lg:col-span-2 xl:col-span-1"
+      title="Clientes"
+      description="28441"
+      icon="pi-users"
+      colorIcon="text-cyan-500"
+      colorBoxIcon="bg-cyan-100"
+    />
+
+    <app-card-widget
+      class="col-span-3 lg:col-span-2 xl:col-span-1"
+      title="Materiales"
+      description="152"
+      icon="pi-box"
+      colorIcon="text-purple-500"
+      colorBoxIcon="bg-purple-100"
+    />
+
+    <app-card-widget
+      class="col-span-3 lg:col-span-2 xl:col-span-1"
+      title="Proveedores"
+      description="152"
+      icon="pi-building"
+      colorIcon="text-yellow-500"
+      colorBoxIcon="bg-purple-100"
+    />
+  `,
 })
 export class StatsWidget {}
